Automatically pull 64bit versions
Adding two lines for each file seems rather
annoying. We've also had issues where we do
not pull both 64 and 32bit libs.

This trys to pull 64bit versions of all entries
with *lib* in them. If the 64 bit versions do not
exist, they are not added to the automatically
generated vendor blob makefile.